Convert buttons on call page to classes
- created EndCallButton
- created ToggleAudioButton
- created ToggleVideoButton
